Add files via upload
Porpose of doing this model is when Indian railway coming inside the JSW pant premises, we need to write the wagon number (Code) from the wagon. For this task we have used various OCR techniques (pytesseract, keras ocr,) but easyocr is giveing better accuracy compare to other. Image quality was very low because of that our model accuracy was impected. We have preprocessed our image the we pass to easyocr to get the batter result.